[ref #190] Make References sortable
